requirements.txt
-e file:///Users/mark/dev/agentic-lab/simulations/agent-dialogues
annotated-types==0.7.0
anyio==4.8.0
argparse==1.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.7.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.40
langchain-ollama==0.2.3
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-cli==0.1.73
langgraph-sdk==0.1.53
langsmith==0.3.11
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
ollama==0.4.7
orjson==3.10.15
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
sniffio==1.3.1
tenacity==9.0.0
typing-extensions==4.12.2
urllib3==2.3.0
zstandard==0.23.0
